--- a/header_test.go
+++ b/header_test.go
@@ -111,14 +111,12 @@ func TestRequestHeaderEmptyValueFromString(t *testing.T) {
 }
 
 func TestRequestRawHeaders(t *testing.T) {
-	kvs := "EmptyValue:\r\n" +
-		"host: foobar\r\n" +
+	kvs := "host: foobar\r\n" +
 		"value: b\r\n" +
 		"\r\n"
 	t.Run("normalized", func(t *testing.T) {
 		s := "GET / HTTP/1.1\r\n" + kvs
-		exp := "Emptyvalue:\r\n" +
-			"Host: foobar\r\n" +
+		exp := "Host: foobar\r\n" +
 			"Value: b\r\n" +
 			"\r\n"
 		var h RequestHeader
@@ -129,16 +127,54 @@ func TestRequestRawHeaders(t *testing.T) {
 		if string(h.Host()) != "foobar" {
 			t.Fatalf("unexpected host: %q. Expecting %q", h.Host(), "foobar")
 		}
-		v1 := h.Peek("EmptyValue")
-		if len(v1) > 0 {
-			t.Fatalf("expecting empty value. Got %q", v1)
+		v2 := h.Peek("Value")
+		if !bytes.Equal(v2, []byte{'b'}) {
+			t.Fatalf("expecting non empty value. Got %q", v2)
+		}
+		if raw := h.RawHeaders(); string(raw) != exp {
+			t.Fatalf("expected header %q, got %q", exp, raw)
+		}
+	})
+	t.Run("post", func(t *testing.T) {
+		s := "POST / HTTP/1.1\r\n" + kvs + "body"
+		exp := "Host: foobar\r\n" +
+			"Value: b\r\n" +
+			"\r\n"
+		var h RequestHeader
+		br := bufio.NewReader(bytes.NewBufferString(s))
+		if err := h.Read(br); err != nil {
+			t.Fatalf("unexpected error: %s", err)
+		}
+		if string(h.Host()) != "foobar" {
+			t.Fatalf("unexpected host: %q. Expecting %q", h.Host(), "foobar")
 		}
 		v2 := h.Peek("Value")
 		if !bytes.Equal(v2, []byte{'b'}) {
 			t.Fatalf("expecting non empty value. Got %q", v2)
 		}
 		if raw := h.RawHeaders(); string(raw) != exp {
-			t.Fatalf("unexpected raw header %q, expected %q instead", kvs, raw)
+			t.Fatalf("expected header %q, got %q", exp, raw)
+		}
+	})
+	t.Run("http10", func(t *testing.T) {
+		s := "GET / HTTP/1.0\r\n" + kvs
+		exp := "Host: foobar\r\n" +
+			"Value: b\r\n" +
+			"\r\n"
+		var h RequestHeader
+		br := bufio.NewReader(bytes.NewBufferString(s))
+		if err := h.Read(br); err != nil {
+			t.Fatalf("unexpected error: %s", err)
+		}
+		if string(h.Host()) != "foobar" {
+			t.Fatalf("unexpected host: %q. Expecting %q", h.Host(), "foobar")
+		}
+		v2 := h.Peek("Value")
+		if !bytes.Equal(v2, []byte{'b'}) {
+			t.Fatalf("expecting non empty value. Got %q", v2)
+		}
+		if raw := h.RawHeaders(); string(raw) != exp {
+			t.Fatalf("expected header %q, got %q", exp, raw)
 		}
 	})
 	t.Run("non-normalized", func(t *testing.T) {
@@ -153,16 +189,12 @@ func TestRequestRawHeaders(t *testing.T) {
 		if string(h.Host()) != "" {
 			t.Fatalf("unexpected host: %q. Expecting %q", h.Host(), "")
 		}
-		v1 := h.Peek("EmptyValue")
-		if len(v1) > 0 {
-			t.Fatalf("expecting empty value. Got %q", v1)
-		}
 		v2 := h.Peek("value")
 		if !bytes.Equal(v2, []byte{'b'}) {
 			t.Fatalf("expecting non empty value. Got %q", v2)
 		}
 		if raw := h.RawHeaders(); string(raw) != exp {
-			t.Fatalf("unexpected raw header %q, expected %q instead", kvs, raw)
+			t.Fatalf("expected header %q, got %q", exp, raw)
 		}
 	})
 	t.Run("no-kvs", func(t *testing.T) {
@@ -182,7 +214,7 @@ func TestRequestRawHeaders(t *testing.T) {
 			t.Fatalf("expecting empty value. Got %q", v1)
 		}
 		if raw := h.RawHeaders(); string(raw) != exp {
-			t.Fatalf("unexpected raw header %q, expected %q instead", kvs, raw)
+			t.Fatalf("expected header %q, got %q", exp, raw)
 		}
 	})
 }