I'm sorry for hassle, but apparently fhttp has fast path header parser for HEAD and GET and `rawHeaders` isn't properly set for other methods. This PR is a fix to #27.

- L1623 now assigns `rawHeaders` to a proper sub-slice for methods with possible body.
- parseHeaders() wasn't returning proper "length parsed", this is fixed by adding internal accumulator in `headerScanner` to properly return parsed length.
- added tests to cover POST and HTTP1.0 cases.

Admittedly this affects global hot path, but any difference is lost in noise:
```sh
alexv@dont-panic ~/s/g/i/fasthttp> benchstat nocounter.res counter.res
name                       old time/op  new time/op  delta
ServerGet10KReqPerConn-8    627ns ± 1%   635ns ± 2%  +1.33%  (p=0.026 n=9+10)
ServerPost10KReqPerConn-8   702ns ± 9%   699ns ± 6%    ~     (p=0.825 n=10+9)
```

Intuitively, one int counter doesn't change much.