@illotum 2018-02-27
I'm observing odd behaviour with non-trivial binary bodies in my fork, where rawHeaders is taken from right from the body and is consequently parsed into (wrong) headers.

Please consider this PR work in progress until I'm able to identify and fix the root cause of this issue.