@illotum 2018-02-27
While this fixed the issue, the performance implications are unfortunately not trivial:
```
name                       old time/op  new time/op  delta
ServerGet10KReqPerConn-8    286ns ± 7%   282ns ±13%     ~     (p=0.424 n=10+10)
ServerPost10KReqPerConn-8   331ns ±14%   390ns ±10%  +17.94%  (p=0.000 n=10+10)
```

@erikdubbelboer For my particular use case this loss is well worth it, but I understand if you would like to postpone the merge.

Ideas how we can work around Peek's sliding window for this particular code path are definitely welcome.