@erikdubbelboer 2018-02-27
How are you benchmarking this? I wouldn't expect a big difference seeing as `rawHeaders` is reused so it only needs to copy the raw header bytes.