@illotum 2018-02-27
Fair point! I'm comparing
```sh
go test -bench='BenchmarkServer(Post|Get)10KReq' -count=10
```
between my branch and your master. Laptop unfortunately.

---
Here's a rerun, this time with lower difference, but also lower confidence.
```
name                       old time/op  new time/op  delta
ServerPost10KReqPerConn-8   319ns ± 3%   349ns ±13%  +9.60%  (p=0.000 n=10+10)
```