@erikdubbelboer 2018-02-28
Could it be that you're mixing up old an new? On my benchmarks the new code always seems to be faster:
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkServerGet10KReqPerConn-8      278           263           -5.40%
BenchmarkServerPost10KReqPerConn-8     350           329           -6.00%

benchmark                              old allocs     new allocs     delta
BenchmarkServerGet10KReqPerConn-8      0              0              +0.00%
BenchmarkServerPost10KReqPerConn-8     0              0              +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkServerGet10KReqPerConn-8      0             0             +0.00%
BenchmarkServerPost10KReqPerConn-8     0             0             +0.00%
```
I'm using Go version go1.10 darwin/amd64.