@illotum 2018-03-01
Funny. No, I'm not mixing it, but this is also not the first time laptop benchmarks fail me. It's Go v1.10 on darwin as well.

I'll post stats from linux/amd64 @ ec2 later today.