@illotum 2018-03-02
```sh
~> benchstat --geomean fhttp postfix  # 10 runs, linux/amd4 @ c4.x8large
name                        old time/op  new time/op  delta
ServerGet10KReqPerConn-36    610ns ± 8%   624ns ± 9%    ~     (p=0.289 n=10+10)
ServerPost10KReqPerConn-36   654ns ±12%   637ns ± 9%    ~     (p=0.403 n=10+10)
[Geo mean]                   632ns        630ns       -0.29%
```
I'm getting convinced any real difference will be noticeable only on the most stable setups. I feel comfortable merging it at this point.