@erikdubbelboer 2018-03-03
On my Ubuntu server with 1.10 the new code is still faster.

I also think it’s ready but just to be sure I just put it live on one of our servers to test. If nothing strange happens I’ll merge it tomorrow.